COMP: Only run auditwheel when building for x64
auditwheel currently does not support cross-compilation.

+++ scripts/internal/manylinux-build-wheels.sh
@@ -156,13 +156,15 @@ for PYBIN in "${PYBINARIES[@]}"; do
 
 done
 
-/opt/python/cp37-cp37m/bin/pip3 install auditwheel wheel
-# Since there are no external shared libraries to bundle into the wheels
-# this step will fixup the wheel switching from 'linux' to 'manylinux2014' tag
-for whl in dist/itk_*linux_$(uname -p).whl; do
-    /opt/python/cp37-cp37m/bin/auditwheel repair --plat manylinux2014_x86_64 ${whl} -w /work/dist/
-    rm ${whl}
-done
+if test "${ARCH}" == "x64"; then
+  /opt/python/cp37-cp37m/bin/pip3 install auditwheel wheel
+  # Since there are no external shared libraries to bundle into the wheels
+  # this step will fixup the wheel switching from 'linux' to 'manylinux2014' tag
+  for whl in dist/itk_*linux_$(uname -p).whl; do
+      /opt/python/cp37-cp37m/bin/auditwheel repair --plat manylinux2014_x86_64 ${whl} -w /work/dist/
+      rm ${whl}
+  done
+fi
 for itk_wheel in dist/itk-*linux*.whl; do
   mv ${itk_wheel} ${itk_wheel/linux/manylinux2014}
 done
